# Starlark: intern dot-expression RHS identifiers globally