Indentifiers are interned in two places:
* when parsing dot-expression RHS
* in `CallUtils` map keys

Thus when evaluating dot expression (`foo.bar`), `bar` is interned
and lookup is faster: `String.equals` performs the identity comparison
before falling back to the comparison of contents.

This PR does interning conservatively: it only interns identifiers
from RHS of dot expression, but no other identifiers.

Other identifiers, and even string literals could be interned to
get performance gain. Consider this example:

```
def my_cc_library_wrapper(**kwargs):
  internal = kwargs.pop("is_internal_library", False)
  if internal:
    local_defs = kwargs.pop("local_defines", []) + ["-DINTERNAL"]
    kwargs["local_defines"] = local_defs
  cc_library(**kwargs)

my_cc_library_wrapper(..., is_internal_library = True)
```

Here we lookup by literal "is_internal_library" which is specified
as a named argument "is_internal_library" in another file: if both
strings were interned, lookup would be faster.

But interning everything may have dire performance implications.
Can be done later.

From my experiments running profiler, attribute lookup is one of
significant bottlenecks of the Starlark interpreter.

Benchmark:

```
def test():
    for _ in range(50000000):
        "abc".startswith
        [].append
        "cde".endswith
        [].extend
        {}.pop
        "fg".rsplit
        [].pop

test()
```

Output is:

```
A: N=27 r=13.546+-0.785 se=0.153
B: N=27 r=11.838+-0.965 se=0.189
B/A: 0.874
```

Or with include benchmark (best of 10 runs 10 seconds each):

```
benchmark                   ops     cpu/op    wall/op   steps/op   alloc/op
# before
bench_dot_expr_native   33554431      353ns      351ns         30       529B
# after
bench_dot_expr_native   67108863      298ns      297ns         30       288B
```