@aiuto 2021-01-06
Is this still relevent?  That is, did the compiler change make it obsolete?