@alandonovan 2021-01-06
The compiler work is still in progress. (I have implemented three different approaches to the compiler and runtime and, demoralizingly, none is yet sufficiently faster than the baseline to justify its complexity.)