@stepancheg 2021-01-06
> none is yet sufficiently faster than the baseline to justify its complexity

What is sufficiently faster by your definition?

My last attempt to implement bytecode was 20% improvement IIRC. That would be enough for us, because it would make the interpreter faster than Python for our use case.

Additionally,
* enabling bytecode interpreter may unlock certain optimizations which might be inappropriate for doing on AST. Like for example, faster Starlark function calls by performing "linking" caller arguments to parameters (`foo(bar=baz, qux=quux)` can be replaced with `foo(baz, quux)` for global functions)
* there are still several missing optimizations in current implementation, for example, indexed predeclared and universe variables, implementing them will make bytecode interpreter speedup better