@stepancheg 2021-02-05
Added another commit on top: inline function parameter names and call argument names to speed up code with many calls with named arguments.

```
A: n=16 mean=9.384 std=0.553 se=0.142 min=8.471 med=9.313
B: n=16 mean=8.903 std=0.559 se=0.144 min=8.148 med=8.906
B/A: 0.949 0.905..0.995 (95% conf)
```

```
def foo(a, bb, c):
  pass

def test():
  for i in range(10):
    print(i)
    for j in range(5000000):
      foo(a=10, bb="rr", c=[])

test()
```