@brandjon 2022-01-11
Note that string interning was changed by 9bf5a14850685c5b0c6231a47809701528d58af1 to intern all idents together, so that'll affect the before/after benchmarks if this change is rebased on head. For instance, that should mean that param names and keyword arg names are already interned together.

There may still be some advantage in further unifying interning across the Starlark interpreter, so ident interning merges with string literal interning (which is currently in the parser since it handles the `'a' + 'b'` -> `'ab'` static optimization) and also with CallUtils' method name lookup. I think in 9bf5a14850685c5b0c6231a47809701528d58af1 I simply considered that kind of experimentation out of scope.